requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
drf-spectacular==0.28.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
cryptography==45.0.5
django-split-settings==1.3.2
redis==6.2.0
python-jose==3.5.0
uvicorn==0.35.0
